Create Calculator-menu-with-radio
Today I learned how to do a menu on web2py, with the help from stackoverflow for my app, when I tried to copy from an e-commerce that I saw on web. Soon I will post here how I improved the menu, because some combinations are impossible to my product, I want it to have a dynamic look and with a different design.